naprrql with napcomp
new release with napcomp, clean-up of old naprr, and unified build scripts.
