+++ app/napval/public/index.html
@@ -2,6 +2,7 @@
 <html>
 
 <head>
+    <title>NIAS - NAPLAN Registration Validation</title>
     <!--Import Google Icon Font-->
     <link href="http://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <!-- jquery support -->
