+++ build_sms.sh
@@ -12,6 +12,7 @@ do_clear() {
 }
 
 do_build() {
+	echo "Building SMS..."
 	mkdir -p $OUTPUT
 	cd $CWD
 	cd ./app/sms
@@ -65,7 +66,7 @@ build_mac64() {
 	do_build
 	#do_upx
 	do_shells
-	do_zip
+	# do_zip
 	echo "...all Mac binaries built..."
 }
 
@@ -84,7 +85,7 @@ build_windows64() {
 	do_build
 	#do_upx
 	do_bats
-	do_zip
+	# do_zip
 	echo "...all Windows64 binaries built..."
 }
 
@@ -102,7 +103,7 @@ build_windows32() {
 	do_build
 	#do_upx
 	do_bats
-	do_zip
+	# do_zip
 	echo "...all Windows32 binaries built..."
 }
 
@@ -120,7 +121,7 @@ build_linux64() {
 	do_build
 	#do_goupx
 	do_shells
-	do_zip
+	# do_zip
 	echo "...all Linux64 binaries built..."
 }
 
@@ -138,7 +139,7 @@ build_linux32() {
 	do_build
 	#do_goupx
 	do_shells
-	do_zip
+	# do_zip
 	echo "...all Linux32 binaries built..."
 }
